Had to add casts to the first two fields in the BitmapFileHeader class to write to the byteArrayBuilder correctly.

+++ src/graphics/bitmaps/file_header.cpp
@@ -1,5 +1,6 @@
 #include "src/graphics/bitmaps/file_header.h"
 #include "src/common/byte_array.h"
+#include <cstdint>
 
 BitmapFileHeader::BitmapFileHeader(int sizeOfPixelArray)
 : BitmapHeader {sizeOfPixelArray}
@@ -15,8 +16,8 @@ char* BitmapFileHeader::toBytes()
 {
   ByteArrayBuilder byteArrayBuilder;
 
-  byteArrayBuilder.addValue(this->bitmapSignatureBytes[0]);
-  byteArrayBuilder.addValue(this->bitmapSignatureBytes[1]);
+  byteArrayBuilder.addValue((uint8_t) this->bitmapSignatureBytes[0]);
+  byteArrayBuilder.addValue((uint8_t) this->bitmapSignatureBytes[1]);
   byteArrayBuilder.addValue(this->sizeOfBitmapFile);
   byteArrayBuilder.addValue(this->reservedBytes);
   byteArrayBuilder.addValue(this->pixelDataOffset);
